refactor: restructure folders & replaced default exports for named exports
