--- src/app/(app)/home/_components/to-do/index.ts
@@ -1,4 +1,0 @@
-export * from './to-do-container';
-export * from './to-do-header';
-export * from './to-do-list';
-export * from './to-do-node';
