+++ src/app/(landing)/(auth)/_components/auth-form/auth-container.tsx
@@ -1,9 +1,10 @@
 'use client';
 
 import { usePathname } from 'next/navigation';
-import { SignInForm, SignUpForm } from './components';
+import { SignInForm } from './sign-in-form';
+import { SignUpForm } from './sign-up-form';
 
-export const AuthForm = () => {
+export const AuthContainer = () => {
   const pathname = usePathname();
 
   return pathname === '/sign-in' ? <SignInForm /> : <SignUpForm />;
