--- src/app/(landing)/(auth)/_components/auth-form/components/index.ts
@@ -1,3 +1,0 @@
-export * from './sign-in-form';
-export * from './sign-up-form';
-export * from './password-input';
