+++ src/app/(landing)/(auth)/_components/auth-form/index.ts
@@ -1,1 +1,4 @@
-export * from './auth-form';
+export * from './auth-container';
+export * from './password-input';
+export * from './sign-in-form';
+export * from './sign-up-form';
