+++ src/app/(landing)/(auth)/_components/auth-form/sign-up-form.tsx
@@ -11,7 +11,7 @@ import {
 } from '@/components/ui/form';
 import { Input } from '@/components/ui/input';
 import { Button } from '@/components/ui/button';
-import PasswordInput from './password-input';
+import { PasswordInput } from './password-input';
 import { signUpSchema } from '@/schemas';
 import { register } from '@/lib/actions/auth.actions';
