+++ src/app/(landing)/(auth)/sign-in/page.tsx
@@ -1,9 +1,9 @@
-import { AuthForm } from '../_components/auth-form';
+import { AuthContainer } from '../_components/auth-form';
 
 const Page = () => {
   return (
     <div className="my-20">
-      <AuthForm />
+      <AuthContainer />
     </div>
   );
 };
