+++ src/app/(landing)/_components/nav/index.ts
